win32 warning

+++ libraries/repro-mysql/src/mysql-api.cpp
@@ -160,7 +160,7 @@ MYSQL_FIELD* statement::field( size_t i ) const
 	if ( i < 0 || i >= column_count_ )
 		throw repro::Ex("invalid field!");
 
-	return mysql_fetch_field_direct(prepare_meta_result_.get(), i);
+	return mysql_fetch_field_direct(prepare_meta_result_.get(), (unsigned long) i);
 }
 
 size_t statement::param_count() const {
